refactor : 프로젝트 환경 분리를 위한 코드
