+++ src/main/java/com/ssafy/ltw/global/util/db/DBProperties.java
@@ -1,0 +1,8 @@
+package com.ssafy.ltw.global.util.db;
+
+public interface DBProperties {
+    String getDriver();
+    String getURL();
+    String getDBId();
+    String getDBPwd();
+}
